master / Первая стабильная сборка 	- Первая полностью рабочая версия программы

+++ lab1/write_file.h
@@ -5,6 +5,13 @@
 #ifndef OS_WRITE_FILE_H
 #define OS_WRITE_FILE_H
 
-int writeFile(int fileId, char* buffer, int numOfBytes);
+/**
+ * Функция пишет numOfBytes байтов в файл
+ * @param fileId - дескриптор файла
+ * @param buffer - буффер, который необходимо записать
+ * @param numOfBytes - кол-во байт, которые надо записать
+ * @return - код упеха операции
+ */
+int writeFile(int fileId, char *buffer, unsigned long numOfBytes);
 
 #endif //OS_WRITE_FILE_H
